Minor cleanup.
