+++ app/src/main/kotlin/com/thomaskioko/tvmaniac/inject/MainActivityComponent.kt
@@ -3,8 +3,8 @@ package com.thomaskioko.tvmaniac.inject
 import android.app.Activity
 import com.thomaskioko.tvmaniac.MainActivityViewModel
 import com.thomaskioko.tvmaniac.navigation.ComposeNavigationFactory
-import com.thomaskioko.tvmaniac.traktauth.TraktAuthManager
-import com.thomaskioko.tvmaniac.traktauth.inject.TraktAuthManagerComponent
+import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthManager
+import com.thomaskioko.tvmaniac.traktauth.implementation.TraktAuthManagerComponent
 import com.thomaskioko.tvmaniac.util.scope.ActivityScope
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Provides
